Adding the first version
Adding all files of the first version of the sorting program TimSort.

TimSort.c contains the main algorithm. TimSortTest.c is just for testing the basic algorithm. ReadStrings.c is just an assistant for read the names of the testing files. There are some testing files in Tests folder.

If you want to use only a sorting algorithm, you only need TimSort.c and TimSort.h.

if you want to test the sorting algorithm, you need to write the files for testing in the Tests folder, enter in the file MainTest.txt the names of the testing files, and then run through the command line a pre-compiled file with the name falya, containing the names of all the testing files and the output falya, where the test results will be recorded. An example file with test results is provided as Result.txt.

for convenience, a makefile is attached to compile and run the program with the run function).