Added all the necessary functions of the real TimSort.
Added gallop function and balancing function for merging.

A program for generating tests has also been added to test the effectiveness of the algorithm over time.

Now, to use this TimSort function, you just need to include in your TimSort.h file and pass the function a pointer to the beginning of the array, the number of elements in it, the size of each element and the comparison function.

This program is still at the testing and refinement stage, so some commands necessary for debugging are still left in the code.