Added more testing functions
added more features for testing. Now each folder is responsible for its own test.

TestingModule is designed to sort numbers modulo.
TestingSets is created for sorting sets. TestingSetsTwo differs from TestingSets in that there the sorting functions are reported addresses on rows with numbers, and not one row with all the numbers in a row. The second one works faster for data where the length of the set is greater than 8 (sizeof(char*)).

If you want to test the sorting according to any of the proposed principles, just go to the desired folder and then run everything in the same way through the make functions.