Improved timrost functions.
Now the balancing function uses a loop, not a recursion

In the merge function, acceleration always works, as my personal research has shown, it speeds up the program somewhat.

The system of working through makefile has been slightly improved. Now it is one for all folders.

As research has shown, my timsort runs about 3 times slower than qsort on random data. But the mail sorted 8-10 times faster.

I hope that this function is working properly now, but I will continue to test it to get rid of errors for sure. It is also possible that I will still optimize it.