Removed unnecessary checks (comparisons)
Now the merge function makes fewer checks by entering an additional variable that remembers the result of the previous check. This speeds up the program by about 7%.

Such an error was found in the testing program that compares timsort and qsort. Before that, there was a problem that the timsort function took an unsorted array, sorted it, and passed the already sorted array to the qsort function. Because of this, their comparison was incorrect.

According to recent measurements, in comparison with the qsort function, the timsort function spends about 5-10% more time on random data, but still about 10 times faster on already sorted ones.